Improve hpa monitor (#63)
* update application URL and hpq query syntax

The web application runs on port 8080, the original URL did not reflect that. It has been updated.
The kubectl get hpa command has the -w flag added to continually query the hpa resource status
without the student having to do so manually.

* update application URL and hpq query syntax

The web application runs on port 8080, the original URL did not reflect that. It has been updated.
The kubectl get hpa command has the -w flag added to continually query the hpa resource status
without the student having to do so manually.