Updated wording to clarify purpose of chapter to for a student who may be newer to Kubernetes concepts
The existing wording was not 100% explicit on where service discovery comes into play for a user
who may not be as experienced with Kubernetes. I called out specifically that service labels are
used in place of static references such as hostnames or IP addresses for the dependent service (webapp)
to communicate with name and greeter.